+++ src/Admin/Authorization/ChiefUserProvider.php
@@ -7,6 +7,7 @@ namespace Thinktomorrow\Chief\Admin\Authorization;
 use Illuminate\Auth\EloquentUserProvider;
 use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Contracts\Auth\UserProvider;
+use Thinktomorrow\Chief\Admin\Users\User;
 
 class ChiefUserProvider extends EloquentUserProvider implements UserProvider
 {
@@ -15,9 +16,20 @@ class ChiefUserProvider extends EloquentUserProvider implements UserProvider
         return parent::newModelQuery($model);
     }
 
+    public function retrieveByToken($identifier, $token)
+    {
+        $user = parent::retrieveByToken($identifier, $token);
+
+        if (! $user || ($user instanceof User && ! $user->isEnabled())) {
+            return null;
+        }
+
+        return $user;
+    }
+
     public function validateCredentials(Authenticatable $user, array $credentials)
     {
-        if (! $user->isEnabled()) {
+        if ($user instanceof User && ! $user->isEnabled()) {
             return false;
         }